requirements.txt
attrs==23.1.0
cffi==1.15.1
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.3
Django==4.2.4
django-crispy-forms==2.0
django-colorfield==0.11.0
django-leaflet == 0.29.1 
fasteners==0.14.1
gast==0.5.2
gunicorn==21.2.0 
html5lib==1.1
jsonfield==3.1.0
lxml==4.8.0
matplotlib==3.5.1
msgpack==1.0.5
numpy==1.21.5
pillow==9.0.1
psycopg2-binary==2.9.6
psycopg2==2.9.9
pycparser==2.21
python-dateutil==2.8.2
pytz==2022.1
scipy==1.8.0
sympy==1.9
tzdata==2023.3